adk-study-planner/webapp/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
